refactor(common): align timeseries market data layers
