+++ src/libs/portfolio-common/portfolio_common/domain/market_data/timeseries.py
