+++ src/libs/portfolio-common/portfolio_common/ports/market_data/__init__.py
@@ -1,0 +1,1 @@
+"""Ports for shared market-data capabilities."""
